apis/bitcoin-extractor/src/requirements.txt
functions-framework==3.0.0
requests
pandas
google-cloud-bigquery